pyproject.toml
[project]
name = "autopolio"
version = "1.0.0"
description = "Portfolio/Resume automation platform with GitHub analysis and LLM-powered summaries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy>=2.0.35",
    "aiosqlite>=0.20.0",

    # Validation and settings
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "email-validator>=2.0.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.2",

    # GitHub API
    "PyGithub>=2.4.0",

    # LLM
    "openai>=1.51.0",
    "anthropic>=0.34.2",
    "google-generativeai>=0.8.0",

    # Document generation
    "python-docx>=1.1.2",
    "reportlab>=4.2.2",
    "PyPDF2>=3.0.1",
    "markdown>=3.7",

    # Encryption
    "cryptography>=43.0.1",

    # Background tasks
    "celery>=5.4.0",
    "redis>=5.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
